Round every product before its sum, as V8 and Rust do
Go may fuse x*y + z into one instruction on arm64, rounding once where
amd64, V8 and Rust round twice. MakeTime and MakeDate, ported from V8
without the conversions go-quickjs's own Date had kept, did so on a Mac:
Date.UTC(1970, 0, 213503982336, 0, 0, 0, -18446744073709552000) was
34448384 there and 34447360 in Node. Date.parse's offset, the Chinese
calendar's day start, and temporal_rs's exact division, whose one
deliberate math.FMA is its point, could be fused the same way. An explicit
float64 conversion now rounds each product first.

TestNoFusedArithmetic compiles the module for arm64 and fails on any fused
instruction a line does not ask for with math.FMA, so that a machine that
never fuses catches one. HostLocale's comment no longer says it is Node's
default locale: V8 makes ICU's fallback en_US_POSIX "en-US".

+++ calchinese.go
@@ -54,7 +54,9 @@ func koreaOffset(ms float64) float64 {
 // daysToMillis is the UTC moment the zone's day starts, the day counted
 // from 1970.
 func (c chineseReckoner) daysToMillis(days float64) float64 {
-	millis := days * astroDayMS
+	// Rounded before the difference, as ICU's two statements are, where an
+	// arm64 build would fuse them.
+	millis := float64(days * astroDayMS)
 	return millis - c.offset(millis)
 }
